+++ serialization-msgpack/src/commonMain/kotlin/com.ensarsarajcic.kotlinx.serialization.msgpack/internal/MsgPackDecoder.kt
@@ -1,6 +1,7 @@
 package com.ensarsarajcic.kotlinx.serialization.msgpack.internal
 
 import com.ensarsarajcic.kotlinx.serialization.msgpack.MsgPackConfiguration
+import com.ensarsarajcic.kotlinx.serialization.msgpack.exceptions.MsgPackSerializationException
 import com.ensarsarajcic.kotlinx.serialization.msgpack.stream.MsgPackDataInputBuffer
 import com.ensarsarajcic.kotlinx.serialization.msgpack.types.MsgPackType
 import com.ensarsarajcic.kotlinx.serialization.msgpack.utils.joinToNumber
@@ -116,7 +117,7 @@ internal class BasicMsgPackDecoder(
                         if (configuration.preventOverflows) {
                             val number = dataBuffer.takeNext(4).joinToNumber<Long>()
                             if (number !in Int.MIN_VALUE..Int.MAX_VALUE) {
-                                throw TODO("Overflow error")
+                                throw MsgPackSerializationException.overflowError(dataBuffer)
                             } else {
                                 number.toInt()
                             }
@@ -125,7 +126,7 @@ internal class BasicMsgPackDecoder(
                         }
                     }
                     else -> {
-                        throw TODO("Add a more descriptive error when wrong type is found!")
+                        throw MsgPackSerializationException.deserialization(dataBuffer, "Unknown array type: $next")
                     }
                 }
 
@@ -137,7 +138,7 @@ internal class BasicMsgPackDecoder(
                         if (configuration.preventOverflows) {
                             val number = dataBuffer.takeNext(4).joinToNumber<Long>()
                             if (number !in Int.MIN_VALUE..Int.MAX_VALUE) {
-                                throw TODO("Overflow error")
+                                throw MsgPackSerializationException.overflowError(dataBuffer)
                             } else {
                                 number.toInt()
                             }
@@ -146,12 +147,12 @@ internal class BasicMsgPackDecoder(
                         }
                     }
                     else -> {
-                        throw TODO("Add a more descriptive error when wrong type is found!")
+                        throw MsgPackSerializationException.deserialization(dataBuffer, "Unknown map type: $next")
                     }
                 }
 
             else -> {
-                TODO("Unsupported collection")
+                throw MsgPackSerializationException.deserialization(dataBuffer, "Unsupported collection: ${descriptor.kind}")
             }
         }
     }
@@ -170,9 +171,8 @@ internal class BasicMsgPackDecoder(
                 return ExtensionTypeDecoder(this)
             }
             // Handle extension types as arrays
-            decodeCollectionSize(descriptor)
-            return ClassMsgPackDecoder(this)
-            // TODO compare with descriptor.elementsCount
+            val size = decodeCollectionSize(descriptor)
+            return ClassMsgPackDecoder(this, size)
         }
         return this
     }
@@ -189,14 +189,15 @@ internal class MsgPackDecoder(
 }
 
 internal class ClassMsgPackDecoder(
-    private val basicMsgPackDecoder: BasicMsgPackDecoder
+    private val basicMsgPackDecoder: BasicMsgPackDecoder,
+    private val size: Int
 ) : Decoder by basicMsgPackDecoder, CompositeDecoder by basicMsgPackDecoder, MsgPackTypeDecoder by basicMsgPackDecoder {
     override val serializersModule: SerializersModule = basicMsgPackDecoder.serializersModule
 
     private var decodedElements = 0
 
     override fun decodeElementIndex(descriptor: SerialDescriptor): Int {
-        if (decodedElements >= descriptor.elementsCount) return CompositeDecoder.DECODE_DONE
+        if (decodedElements >= size) return CompositeDecoder.DECODE_DONE
         val result = basicMsgPackDecoder.decodeElementIndex(descriptor)
         if (result != CompositeDecoder.DECODE_DONE) decodedElements++
         return result
@@ -223,7 +224,7 @@ internal class ExtensionTypeDecoder(
             val byte = dataBuffer.requireNextByte()
             bytesRead++
             if (!MsgPackType.Ext.isExt(byte)) {
-                throw TODO("Unexpected byte")
+                throw MsgPackSerializationException.deserialization(dataBuffer, "Unexpected byte: $byte. Expected extension type byte!")
             }
             type = byte
             if (MsgPackType.Ext.SIZES.containsKey(type)) {
@@ -243,7 +244,7 @@ internal class ExtensionTypeDecoder(
             typeId = byte
             typeId!!
         } else {
-            throw TODO("Handle?")
+            throw AssertionError()
         }
     }
     override fun decodeElementIndex(descriptor: SerialDescriptor): Int {