+++ serialization-msgpack/src/commonMain/kotlin/com.ensarsarajcic.kotlinx.serialization.msgpack/internal/MsgPacker.kt
@@ -1,5 +1,6 @@
 package com.ensarsarajcic.kotlinx.serialization.msgpack.internal
 
+import com.ensarsarajcic.kotlinx.serialization.msgpack.exceptions.MsgPackSerializationException
 import com.ensarsarajcic.kotlinx.serialization.msgpack.types.MsgPackType
 import com.ensarsarajcic.kotlinx.serialization.msgpack.utils.splitToByteArray
 
@@ -113,7 +114,7 @@ internal class BasicMsgPacker : MsgPacker {
             bytes.size <= MsgPackType.String.MAX_STR32_LENGTH -> {
                 byteArrayOf(MsgPackType.String.STR32) + bytes.size.toInt().splitToByteArray()
             }
-            else -> TODO("TOO LONG STRING")
+            else -> throw MsgPackSerializationException.packingError("String too long. Byte size: ${bytes.size}. Max size: ${MsgPackType.String.MAX_STR32_LENGTH}")
         }
         return prefix + bytes
     }
@@ -129,7 +130,7 @@ internal class BasicMsgPacker : MsgPacker {
             value.size <= MsgPackType.Bin.MAX_BIN32_LENGTH -> {
                 byteArrayOf(MsgPackType.Bin.BIN32) + value.size.toInt().splitToByteArray()
             }
-            else -> TODO("TOO LONG STRING")
+            else -> throw MsgPackSerializationException.packingError("Byte array too long. Byte size: ${value.size}. Max size: ${MsgPackType.Bin.MAX_BIN32_LENGTH}")
         }
         return prefix + value
     }